+++ lib/recording_studio_terms_and_conditions/sample_terms.rb
@@ -2,7 +2,7 @@
 
 module RecordingStudioTermsAndConditions
   module SampleTerms
-    TITLE = "Studio Terms"
+    TITLE = "Terms and Conditions"
 
     BODY = <<~HTML
       <h2>Welcome</h2>